% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callmix.labelling.R
\name{callmix.labelling}
\alias{callmix.labelling}
\title{Classification of positions as fishing or steaming, using speed data.}
\usage{
callmix.labelling(x, nclust, algo = "depmix")
}
\arguments{
\item{x}{a vector of the speed data}

\item{nclust}{the number of clusters}

\item{algo}{'depmix' or 'mclust'}
}
\value{
A list of 3 elements:
`states` A vector of the classified fishing or steaming states.
`loglik` Final log-likelihood estimate of the EM algorithm.
`parameters` A list of parameters dependent on the model.
}
\description{
Classification of positions as fishing or steaming, using speed data.
}
\details{
By using the methods provided in the \code{algo} argument, model-based clustering is performed, with the specified number of clusters. The clusters are then assigned to fishing of steaming using the following labelling scheme: Fishing corresponds to the combination of components that has a joint empirical variance smaller than that of the low-speed states estimated in a 2-component model. If there is more than one combination that reduces the variance, we choose the one with a joint empirical mean closest to that of the low-speed states estimates in a 2-component model.
}
\examples{
data("DanishTrips")
res <- callmix.labelling(DanishTrips$speed,3)
}
