% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.circular.R
\name{classify.circular}
\alias{classify.circular}
\title{Classification of positions as fishing or steaming, using speed and hour of the day data.}
\usage{
classify.circular(speed, time, method, diag = F)
}
\arguments{
\item{speed}{a vector of speed data.}

\item{time}{a vector of hour of the day data.}

\item{method}{'timeclust' or 'circlust'}

\item{diag}{a boolean to be passed to  'circlust'}
}
\value{
A list of 3 elements: 
`classification` A vector of the classified fishing or steaming states.
`loglik` Final log-likelihood estimate of the EM algorithm. 
`parameters` A list of parameters dependent on the model.
}
\description{
Classification of positions as fishing or steaming, using speed and hour of the day data, with model-based clustering.
}
\details{
The classification is performed in two stages. First, 
a clustering with two clusters is performed on the speed data. 
The cluster with the high speed mean is classified as steaming. 
The rest of the positions are used as an input for the `timeclust``
or `circlust` methods. From the resulting classification, 
the positions that have a time mean closer to 9 are classified as fishing and the rest as steaming.
}
\examples{
dates <- as.timeDate(DanishTrips$DATES)
time <- hour(dates)+minute(dates)/60
res <- classify.circular(DanishTrips$speed, time, 'timeclust')
misclassification <- function(v, s)
{
  d <- 0
  for (i in 1:length(v))
    if (v[i] != s[i])
      d <- d+1
    return(d/length(v)*100)
}
misclassification(res$classification, DanishTrips$state_seq)
}
