% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeclust.R
\name{timeclust}
\alias{timeclust}
\title{Model-based clustering for one-dimensional circular data.}
\usage{
timeclust(x)
}
\arguments{
\item{x}{a vector with the hour of the day data.}
}
\value{
\item{loglik }{Final log-likelihood estimate of the EM algorithm.}
\item{parameters }{Parameters inferred by the algorithm:
    \itemize{
      \item pro: Mixing proportion of each distribution.
      \item mean: Means of the two clusters.
      \item sigma: Sigmas of the two clusters.
      \item z: Responsibilities.}}
      \item{classification }{Classification of the datapoints to the clusters.}
\item{iter }{Number of iterations of the algorithm.}
}
\description{
Model-based clustering with two clusters for one-dimensional hour of the day data. Expectation maximization algorithm is implemented that takes into account that the hour of the day data are circular, i.e. 00:00 is the same as 24:00. The circularity is being handled by defining a truncated normal distribution.
}
\examples{
x1 <- rnorm(100, 0, 3)
x1[x1<0] <- x1[x1<0]+24
x2 <- rnorm(80, 9, 2)
x <- c(x1,x2)
x <- x[sample.int(length(x))]
res <- timeclust(x)
cat('mixing proportions:',res$parameters$pro,'\\nmeans:',
    res$parameters$mean,'\\nsigmas:',res$parameters$sigma)
}
